Move feature to own file `mute-users`
