+++ source/content.ts
@@ -105,6 +105,7 @@ import './features/show-asset-download-count';
 import './features/open-issue-to-latest-comment';
 import './features/highest-rated-comment';
 import './features/clean-issue-filters';
+import './features/mute-users';
 
 import './features/scrollable-code-and-blockquote.css';
 import './features/center-reactions-popup.css';
