+++ source/options.html
@@ -43,15 +43,10 @@
 		</label>
 	</p>
 
-	<hr>
-
-	<p>
-			<label>
-				<strong>Muted users</strong>
-				<p>Comments from muted users will be minimized in issue and PR conversations.</p>
-				<textarea name="mutedUsers" rows="2" spellcheck="false"></textarea>
-			</label>
-	</p>
+	<div class="js-muted-users-container extended-options">
+		<span>Users muted by this feature:</span>
+		<textarea name="mutedUsers" rows="2" spellcheck="false"></textarea>
+	</div>
 
 	<hr>
